@@ -1,6 +1,6 @@
 name: flutter_cupertino_date_picker
 description: Flutter cupertino date picker. Display DatePicker in iOS style. Support both Android and iOS platform.
-version: 1.0.26+2
+version: 2.0.0+3
 author: Dylan Wu <dylanwuzh@gmail.com>
 homepage: https://github.com/wuzhendev/flutter-cupertino-date-picker
 
@@ -11,5 +11,5 @@ dependencies:
 flutter:
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
-  flutter: ">=0.4.1 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.22.5"
