@@ -18,7 +18,7 @@ const List<int> _solarMonthsOf31Days = const <int>[1, 3, 5, 7, 8, 10, 12];
 /// @since 2019-05-10
 class DatePickerWidget extends StatefulWidget {
   DatePickerWidget({
-    Key key,
+    Key? key,
     this.onMonthChangeStartWithFirstDate,
     this.minDateTime,
     this.maxDateTime,
@@ -35,13 +35,13 @@ class DatePickerWidget extends StatefulWidget {
     assert(minTime.compareTo(maxTime) < 0);
   }
 
-  final DateTime minDateTime, maxDateTime, initialDateTime;
-  final String dateFormat;
-  final DateTimePickerLocale locale;
-  final DateTimePickerTheme pickerTheme;
+  final DateTime? minDateTime, maxDateTime, initialDateTime;
+  final String? dateFormat;
+  final DateTimePickerLocale? locale;
+  final DateTimePickerTheme? pickerTheme;
 
-  final DateVoidCallback onCancel;
-  final DateValueCallback onChange, onConfirm;
+  final DateVoidCallback? onCancel;
+  final DateValueCallback? onChange, onConfirm;
   final onMonthChangeStartWithFirstDate;
 
   @override
@@ -50,18 +50,18 @@ class DatePickerWidget extends StatefulWidget {
 }
 
 class _DatePickerWidgetState extends State<DatePickerWidget> {
-  DateTime _minDateTime, _maxDateTime;
-  int _currYear, _currMonth, _currDay;
-  List<int> _yearRange, _monthRange, _dayRange;
-  FixedExtentScrollController _yearScrollCtrl, _monthScrollCtrl, _dayScrollCtrl;
+  DateTime? _minDateTime, _maxDateTime;
+  int? _currYear, _currMonth, _currDay;
+  List<int>? _yearRange, _monthRange, _dayRange;
+  FixedExtentScrollController? _yearScrollCtrl, _monthScrollCtrl, _dayScrollCtrl;
 
-  Map<String, FixedExtentScrollController> _scrollCtrlMap;
-  Map<String, List<int>> _valueRangeMap;
+  late Map<String, FixedExtentScrollController?> _scrollCtrlMap;
+  late Map<String, List<int>?> _valueRangeMap;
 
   bool _isChangeDateRange = false;
 
   _DatePickerWidgetState(
-      DateTime minDateTime, DateTime maxDateTime, DateTime initialDateTime) {
+      DateTime? minDateTime, DateTime? maxDateTime, DateTime? initialDateTime) {
     // handle current selected year、month、day
     DateTime initDateTime = initialDateTime ?? DateTime.now();
     this._currYear = initDateTime.year;
@@ -74,23 +74,23 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
     // limit the range of year
     this._yearRange = _calcYearRange();
-    this._currYear = min(max(_minDateTime.year, _currYear), _maxDateTime.year);
+    this._currYear = min(max(_minDateTime!.year, _currYear!), _maxDateTime!.year);
 
     // limit the range of month
     this._monthRange = _calcMonthRange();
-    this._currMonth = min(max(_monthRange.first, _currMonth), _monthRange.last);
+    this._currMonth = min(max(_monthRange!.first, _currMonth!), _monthRange!.last);
 
     // limit the range of day
     this._dayRange = _calcDayRange();
-    this._currDay = min(max(_dayRange.first, _currDay), _dayRange.last);
+    this._currDay = min(max(_dayRange!.first, _currDay!), _dayRange!.last);
 
     // create scroll controller
     _yearScrollCtrl =
-        FixedExtentScrollController(initialItem: _currYear - _yearRange.first);
+        FixedExtentScrollController(initialItem: _currYear! - _yearRange!.first);
     _monthScrollCtrl = FixedExtentScrollController(
-        initialItem: _currMonth - _monthRange.first);
+        initialItem: _currMonth! - _monthRange!.first);
     _dayScrollCtrl =
-        FixedExtentScrollController(initialItem: _currDay - _dayRange.first);
+        FixedExtentScrollController(initialItem: _currDay! - _dayRange!.first);
 
     _scrollCtrlMap = {
       'y': _yearScrollCtrl,
@@ -113,7 +113,7 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
     Widget datePickerWidget = _renderDatePickerWidget();
 
     // display the title widget
-    if (widget.pickerTheme.title != null || widget.pickerTheme.showTitle) {
+    if (widget.pickerTheme!.title != null || widget.pickerTheme!.showTitle) {
       Widget titleWidget = DatePickerTitleWidget(
         pickerTheme: widget.pickerTheme,
         locale: widget.locale,
@@ -128,7 +128,7 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   /// pressed cancel widget
   void _onPressedCancel() {
     if (widget.onCancel != null) {
-      widget.onCancel();
+      widget.onCancel!();
     }
     Navigator.pop(context);
   }
@@ -136,8 +136,8 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   /// pressed confirm widget
   void _onPressedConfirm() {
     if (widget.onConfirm != null) {
-      DateTime dateTime = DateTime(_currYear, _currMonth, _currDay);
-      widget.onConfirm(dateTime, _calcSelectIndexList());
+      DateTime dateTime = DateTime(_currYear!, _currMonth!, _currDay!);
+      widget.onConfirm!(dateTime, _calcSelectIndexList());
     }
     Navigator.pop(context);
   }
@@ -145,14 +145,14 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   /// notify selected date changed
   void _onSelectedChange() {
     if (widget.onChange != null) {
-      DateTime dateTime = DateTime(_currYear, _currMonth, _currDay);
-      widget.onChange(dateTime, _calcSelectIndexList());
+      DateTime dateTime = DateTime(_currYear!, _currMonth!, _currDay!);
+      widget.onChange!(dateTime, _calcSelectIndexList());
     }
   }
 
   /// find scroll controller by specified format
-  FixedExtentScrollController _findScrollCtrl(String format) {
-    FixedExtentScrollController scrollCtrl;
+  FixedExtentScrollController? _findScrollCtrl(String format) {
+    FixedExtentScrollController? scrollCtrl;
     _scrollCtrlMap.forEach((key, value) {
       if (format.contains(key)) {
         scrollCtrl = value;
@@ -162,8 +162,8 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   }
 
   /// find item value range by specified format
-  List<int> _findPickerItemRange(String format) {
-    List<int> valueRange;
+  List<int>? _findPickerItemRange(String format) {
+    List<int>? valueRange;
     _valueRangeMap.forEach((key, value) {
       if (format.contains(key)) {
         valueRange = value;
@@ -174,11 +174,11 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
   /// render the picker widget of year、month and day
   Widget _renderDatePickerWidget() {
-    List<Widget> pickers = List<Widget>();
+    List<Widget> pickers = [];
     List<String> formatArr =
         DateTimeFormatter.splitDateFormat(widget.dateFormat);
     formatArr.forEach((format) {
-      List<int> valueRange = _findPickerItemRange(format);
+      List<int> valueRange = _findPickerItemRange(format)!;
 
       Widget pickerColumn = _renderDatePickerColumnComponent(
         scrollCtrl: _findScrollCtrl(format),
@@ -201,21 +201,21 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   }
 
   Widget _renderDatePickerColumnComponent({
-    @required FixedExtentScrollController scrollCtrl,
-    @required List<int> valueRange,
-    @required String format,
-    @required ValueChanged<int> valueChanged,
+    required FixedExtentScrollController? scrollCtrl,
+    required List<int> valueRange,
+    required String format,
+    required ValueChanged<int> valueChanged,
   }) {
     return Expanded(
       flex: 1,
       child: Container(
         padding: EdgeInsets.all(8.0),
-        height: widget.pickerTheme.pickerHeight,
-        decoration: BoxDecoration(color: widget.pickerTheme.backgroundColor),
+        height: widget.pickerTheme!.pickerHeight,
+        decoration: BoxDecoration(color: widget.pickerTheme!.backgroundColor),
         child: CupertinoPicker.builder(
-          backgroundColor: widget.pickerTheme.backgroundColor,
+          backgroundColor: widget.pickerTheme!.backgroundColor,
           scrollController: scrollCtrl,
-          itemExtent: widget.pickerTheme.itemHeight,
+          itemExtent: widget.pickerTheme!.itemHeight,
           onSelectedItemChanged: valueChanged,
           childCount: valueRange.last - valueRange.first + 1,
           itemBuilder: (context, index) =>
@@ -227,19 +227,19 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
   Widget _renderDatePickerItemComponent(int value, String format) {
     return Container(
-      height: widget.pickerTheme.itemHeight,
+      height: widget.pickerTheme!.itemHeight,
       alignment: Alignment.center,
       child: Text(
         DateTimeFormatter.formatDateTime(value, format, widget.locale),
         style:
-            widget.pickerTheme.itemTextStyle ?? DATETIME_PICKER_ITEM_TEXT_STYLE,
+            widget.pickerTheme?.itemTextStyle ?? DATETIME_PICKER_ITEM_TEXT_STYLE,
       ),
     );
   }
 
   /// change the selection of year picker
   void _changeYearSelection(int index) {
-    int year = _yearRange.first + index;
+    int year = _yearRange!.first + index;
     if (_currYear != year) {
       _currYear = year;
       _changeDateRange();
@@ -249,7 +249,7 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
   /// change the selection of month picker
   void _changeMonthSelection(int index) {
-    int month = _monthRange.first + index;
+    int month = _monthRange!.first + index;
     if (_currMonth != month) {
       _currMonth = month;
       _changeDateRange();
@@ -259,7 +259,7 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
   /// change the selection of day picker
   void _changeDaySelection(int index) {
-    int dayOfMonth = _dayRange.first + index;
+    int dayOfMonth = _dayRange!.first + index;
     if (_currDay != dayOfMonth) {
       _currDay = dayOfMonth;
       _onSelectedChange();
@@ -274,20 +274,20 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
     _isChangeDateRange = true;
 
     List<int> monthRange = _calcMonthRange();
-    bool monthRangeChanged = _monthRange.first != monthRange.first ||
-        _monthRange.last != monthRange.last;
+    bool monthRangeChanged = _monthRange!.first != monthRange.first ||
+        _monthRange!.last != monthRange.last;
     if (monthRangeChanged) {
       // selected year changed
-      _currMonth = max(min(_currMonth, monthRange.last), monthRange.first);
+      _currMonth = max(min(_currMonth!, monthRange.last), monthRange.first);
     }
 
     List<int> dayRange = _calcDayRange();
     bool dayRangeChanged =
-        _dayRange.first != dayRange.first || _dayRange.last != dayRange.last;
+        _dayRange!.first != dayRange.first || _dayRange!.last != dayRange.last;
     if (dayRangeChanged) {
       // day range changed, need limit the value of selected day
       if (!widget.onMonthChangeStartWithFirstDate) {
-        max(min(_currDay, dayRange.last), dayRange.first);
+        max(min(_currDay!, dayRange.last), dayRange.first);
       } else {
         _currDay = dayRange.first;
       }
@@ -303,19 +303,19 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
     if (monthRangeChanged) {
       // CupertinoPicker refresh data not working (https://github.com/flutter/flutter/issues/22999)
-      int currMonth = _currMonth;
-      _monthScrollCtrl.jumpToItem(monthRange.last - monthRange.first);
+      int currMonth = _currMonth!;
+      _monthScrollCtrl!.jumpToItem(monthRange.last - monthRange.first);
       if (currMonth < monthRange.last) {
-        _monthScrollCtrl.jumpToItem(currMonth - monthRange.first);
+        _monthScrollCtrl!.jumpToItem(currMonth - monthRange.first);
       }
     }
 
     if (dayRangeChanged) {
       // CupertinoPicker refresh data not working (https://github.com/flutter/flutter/issues/22999)
-      int currDay = _currDay;
-      _dayScrollCtrl.jumpToItem(dayRange.last - dayRange.first);
+      int currDay = _currDay!;
+      _dayScrollCtrl!.jumpToItem(dayRange.last - dayRange.first);
       if (currDay < dayRange.last) {
-        _dayScrollCtrl.jumpToItem(currDay - dayRange.first);
+        _dayScrollCtrl!.jumpToItem(currDay - dayRange.first);
       }
     }
 
@@ -325,7 +325,7 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   /// calculate the count of day in current month
   int _calcDayCountOfMonth() {
     if (_currMonth == 2) {
-      return isLeapYear(_currYear) ? 29 : 28;
+      return isLeapYear(_currYear!) ? 29 : 28;
     } else if (_solarMonthsOf31Days.contains(_currMonth)) {
       return 31;
     }
@@ -339,30 +339,29 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
 
   /// calculate selected index list
   List<int> _calcSelectIndexList() {
-    int yearIndex = _currYear - _minDateTime.year;
-    int monthIndex = _currMonth - _monthRange.first;
-    int dayIndex = _currDay - _dayRange.first;
+    int yearIndex = _currYear! - _minDateTime!.year;
+    int monthIndex = _currMonth! - _monthRange!.first;
+    int dayIndex = _currDay! - _dayRange!.first;
     return [yearIndex, monthIndex, dayIndex];
   }
 
   /// calculate the range of year
-  List<int> _calcYearRange() {
-    print(_minDateTime.year.toString() + ", " + _maxDateTime.year.toString());
-    return [_minDateTime.year, _maxDateTime.year];
+  List<int>? _calcYearRange() {
+    return [_minDateTime?.year??0, _maxDateTime?.year??0];
   }
 
   /// calculate the range of month
   List<int> _calcMonthRange() {
     int minMonth = 1, maxMonth = 12;
-    int minYear = _minDateTime.year;
-    int maxYear = _maxDateTime.year;
+    int? minYear = _minDateTime?.year;
+    int? maxYear = _maxDateTime?.year;
     if (minYear == _currYear) {
       // selected minimum year, limit month range
-      minMonth = _minDateTime.month;
+      minMonth = _minDateTime?.month??1;
     }
     if (maxYear == _currYear) {
       // selected maximum year, limit month range
-      maxMonth = _maxDateTime.month;
+      maxMonth = _maxDateTime?.month??12;
     }
     return [minMonth, maxMonth];
   }
@@ -370,20 +369,20 @@ class _DatePickerWidgetState extends State<DatePickerWidget> {
   /// calculate the range of day
   List<int> _calcDayRange({currMonth}) {
     int minDay = 1, maxDay = _calcDayCountOfMonth();
-    int minYear = _minDateTime.year;
-    int maxYear = _maxDateTime.year;
-    int minMonth = _minDateTime.month;
-    int maxMonth = _maxDateTime.month;
+    int? minYear = _minDateTime?.year;
+    int? maxYear = _maxDateTime?.year;
+    int? minMonth = _minDateTime?.month;
+    int? maxMonth = _maxDateTime?.month;
     if (currMonth == null) {
       currMonth = _currMonth;
     }
     if (minYear == _currYear && minMonth == currMonth) {
       // selected minimum year and month, limit day range
-      minDay = _minDateTime.day;
+      minDay = _minDateTime?.day??1;
     }
     if (maxYear == _currYear && maxMonth == currMonth) {
       // selected maximum year and month, limit day range
-      maxDay = _maxDateTime.day;
+      maxDay = _maxDateTime?.day??_calcDayCountOfMonth();
     }
     return [minDay, maxDay];
   }
