@@ -31,15 +31,15 @@ class DateTimeFormatter {
   }
 
   /// Split date format to array.
-  static List<String> splitDateFormat(String dateFormat,
-      {DateTimePickerMode mode}) {
+  static List<String> splitDateFormat(String? dateFormat,
+      {DateTimePickerMode? mode}) {
     if (dateFormat == null || dateFormat.length == 0) {
       return [];
     }
     List<String> result = dateFormat.split(RegExp(DATE_FORMAT_SEPARATOR));
     if (mode == DateTimePickerMode.datetime) {
       // datetime mode need join day format
-      List<String> temp = List<String>();
+      List<String> temp = [];
       StringBuffer dayFormat = StringBuffer();
       for (int i = 0; i < result.length; i++) {
         String format = result[i];
@@ -71,7 +71,7 @@ class DateTimeFormatter {
 
   /// Format datetime string
   static String formatDateTime(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     if (format == null || format.length == 0) {
       return value.toString();
     }
@@ -112,7 +112,7 @@ class DateTimeFormatter {
 
   /// Format day display
   static String formatDate(
-      DateTime dateTime, String format, DateTimePickerLocale locale) {
+      DateTime dateTime, String format, DateTimePickerLocale? locale) {
     if (format == null || format.length == 0) {
       return dateTime.toString();
     }
@@ -141,7 +141,7 @@ class DateTimeFormatter {
 
   /// format year text
   static String _formatYear(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     if (format.contains('yyyy')) {
       // yyyy: the digit count of year is 4, e.g. 2019
       return format.replaceAll('yyyy', value.toString());
@@ -155,15 +155,15 @@ class DateTimeFormatter {
 
   /// format month text
   static String _formatMonth(
-      int value, String format, DateTimePickerLocale locale) {
-    List<String> months = DatePickerI18n.getLocaleMonths(locale);
+      int value, String format, DateTimePickerLocale? locale) {
+    List<String>? months = DatePickerI18n.getLocaleMonths(locale);
     if (format.contains('MMMM')) {
       // MMMM: the full name of month, e.g. January
-      return format.replaceAll('MMMM', months[value - 1]);
+      return format.replaceAll('MMMM', months![value - 1]);
     } else if (format.contains('MMM')) {
       // MMM: the short name of month, e.g. Jan
       months = DatePickerI18n.getLocaleMonths(locale, false);
-      String month = months[value - 1];
+      String month = months![value - 1];
       return format.replaceAll('MMM', month);
     }
     return _formatNumber(value, format, 'M');
@@ -171,38 +171,38 @@ class DateTimeFormatter {
 
   /// format day text
   static String _formatDay(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     return _formatNumber(value, format, 'd');
   }
 
   /// format week text
   static String _formatWeek(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     if (format.contains('EEEE')) {
       // EEEE: the full name of week, e.g. Monday
-      List<String> weeks = DatePickerI18n.getLocaleWeeks(locale);
+      List<String> weeks = DatePickerI18n.getLocaleWeeks(locale)!;
       return format.replaceAll('EEEE', weeks[value - 1]);
     }
     // EEE: the short name of week, e.g. Mon
-    List<String> weeks = DatePickerI18n.getLocaleWeeks(locale, false);
+    List<String> weeks = DatePickerI18n.getLocaleWeeks(locale, false)!;
     return format.replaceAll(RegExp(r'E+'), weeks[value - 1]);
   }
 
   /// format hour text
   static String _formatHour(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     return _formatNumber(value, format, 'H');
   }
 
   /// format minute text
   static String _formatMinute(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     return _formatNumber(value, format, 'm');
   }
 
   /// format second text
   static String _formatSecond(
-      int value, String format, DateTimePickerLocale locale) {
+      int value, String format, DateTimePickerLocale? locale) {
     return _formatNumber(value, format, 's');
   }
 
