@@ -29,22 +29,22 @@ abstract class _StringsI18n {
   const _StringsI18n();
 
   /// Get the done widget text
-  String getDoneText();
+  String? getDoneText();
 
   /// Get the cancel widget text
-  String getCancelText();
+  String? getCancelText();
 
   /// Get the name of month
   List<String> getMonths();
 
   /// Get the short name of month
-  List<String> getMonthsShort();
+  List<String>? getMonthsShort();
 
   /// Get the full name of week
   List<String> getWeeksFull();
 
   /// Get the short name of week
-  List<String> getWeeksShort();
+  List<String>? getWeeksShort();
 }
 
 enum DateTimePickerLocale {
@@ -154,56 +154,54 @@ const Map<DateTimePickerLocale, _StringsI18n> datePickerI18n = {
 
 class DatePickerI18n {
   /// Get done button text
-  static String getLocaleDone(DateTimePickerLocale locale) {
-    _StringsI18n i18n = datePickerI18n[locale] ??
-        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT];
-    return i18n.getDoneText() ??
-        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getDoneText();
+  static String? getLocaleDone(DateTimePickerLocale? locale) {
+    _StringsI18n i18n = datePickerI18n[locale!] ??
+        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!;
+    return i18n.getDoneText() ?? datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getDoneText();
   }
 
   /// Get cancel button text
-  static String getLocaleCancel(DateTimePickerLocale locale) {
-    _StringsI18n i18n = datePickerI18n[locale] ??
-        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT];
-    return i18n.getCancelText() ??
-        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getCancelText();
+  static String? getLocaleCancel(DateTimePickerLocale? locale) {
+    _StringsI18n i18n = datePickerI18n[locale!] ??
+        datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!;
+    return i18n.getCancelText() ?? datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getCancelText();
   }
 
   /// Get locale month array
-  static List<String> getLocaleMonths(DateTimePickerLocale locale,
+  static List<String>? getLocaleMonths(DateTimePickerLocale? locale,
       [bool isFull = true]) {
-    _StringsI18n i18n = datePickerI18n[locale] ??
+    _StringsI18n? i18n = datePickerI18n[locale!] ??
         datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT];
 
     if (isFull) {
-      List<String> months = i18n.getMonths();
+      List<String> months = i18n!.getMonths();
       if (months != null && months.isNotEmpty) {
         return months;
       }
-      return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getMonths();
+      return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getMonths();
     }
 
-    List<String> months = i18n.getMonthsShort();
+    List<String>? months = i18n!.getMonthsShort();
     if (months != null && months.isNotEmpty && months.length == 12) {
       return months;
     }
-    return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getMonthsShort();
+    return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getMonthsShort();
   }
 
   /// Get locale week array
-  static List<String> getLocaleWeeks(DateTimePickerLocale locale,
+  static List<String>? getLocaleWeeks(DateTimePickerLocale? locale,
       [bool isFull = true]) {
-    _StringsI18n i18n = datePickerI18n[locale] ??
+    _StringsI18n? i18n = datePickerI18n[locale!] ??
         datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT];
     if (isFull) {
-      List<String> weeks = i18n.getWeeksFull();
+      List<String> weeks = i18n!.getWeeksFull();
       if (weeks != null && weeks.isNotEmpty) {
         return weeks;
       }
-      return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getWeeksFull();
+      return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getWeeksFull();
     }
 
-    List<String> weeks = i18n.getWeeksShort();
+    List<String>? weeks = i18n!.getWeeksShort();
     if (weeks != null && weeks.isNotEmpty) {
       return weeks;
     }
@@ -214,6 +212,6 @@ class DatePickerI18n {
           .map((item) => item.substring(0, min(3, item.length)))
           .toList();
     }
-    return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT].getWeeksShort();
+    return datePickerI18n[DATETIME_PICKER_LOCALE_DEFAULT]!.getWeeksShort();
   }
 }
